requirements.txt
django
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.1
django-environ==0.11.2
pre-commit==3.7.0
django-redis==5.4.0
kazoo==2.10.0
